Update plugin information
The most important change here is to point to the BlyncSync version of the plugin, which is actively maintained.  The jins-meme version lacks a package.json, which causes issues on some platforms that will result in an inability to build.

The BlyncSync plugin uses standard naming convention for cordova plugins (cordova-plugin-jins-meme-es)

The plugin name should be JINS MEME ES (ES for Eye-Sensing,) as this is how the product is actually named: not "Jins Meme."