feat(jins-meme): enable background mode data collection
In order to enable background mode data collection we used the cordova-plugin-background-mode plugin in a new Cordova plugin repo, cordova-plugin-jins-meme (vs. cordova-plugin-jins-meme-es.)  We opted to maintain the old repository for backwards-compatibility as we have multiple apps using various versions; this is simplest.

Another unrelated change was made to silence unnecessary console logging.